requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-jose==3.5.0
python-dotenv==1.1.1
pymongo==4.14.1
python-multipart==0.0.20
bcrypt==4.3.0
pytest==8.4.1
pytest-asyncio==1.1.0 
httpx==0.28.1